task6fix

+++ hello_world/6-size.c
@@ -11,11 +11,10 @@ long long int c;
 char d;
 float f;
 
-printf("Size of char: %d byte(s)\n", sizeof(d));
-printf("Size an int: %d byte(s)\n", sizeof(a));
-printf("Size of long int: %d byte(s)\n", sizeof(b));
-printf("Size of long long int: %d byte(s)\n", sizeof(c));
-printf("Size of float: %d byte(s)\n", sizeof(f));
-
+printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(d));
+printf("Size an int: %lu byte(s)\n", (unsigned long)sizeof(a));
+printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(b));
+printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
